1. Created DataGrid component. 2. Created DatePicker component. 3. Refactored Cohort page to use reusable component for MUI inputs: DataGrid, DatePicker
